Progress bar cleanup (#96)
* remove unused build_progress_bar method

* remove noop/none logging as a log format option

* make log_format default to json and remove default_log_format arg

* move progress_bar to submodule

* rename to base_progress_bar

* module rename

* move out json progress bar

* move out tensorboard progress bar

* move out wandb progress wrapper

* split out helpers to utils, avoid circular import

* fix broken init

* remove module rename for progress bar

* move get_precise_epoch to higher level utils

* lint

* remove no progress bar flag

* fix more imports

* cleanup

* remove utils

* add license
