+++ metaseq/dataclass/configs.py
@@ -84,9 +84,6 @@ class CommonConfig(MetaseqDataclass):
     # different jobs. Please append your params to other dataclasses if they
     # were used for a particular purpose or task, such as those dedicated for
     # `distributed training`, `optimization`, etc.
-    no_progress_bar: bool = field(
-        default=False, metadata={"help": "disable progress bar"}
-    )
     log_interval: int = field(
         default=100,
         metadata={
