+++ metaseq_cli/train.py
@@ -241,7 +241,7 @@ def train(
             ),
         )
 
-    progress = progress_bar.progress_bar(
+    progress = progress_bar.get_progress_bar(
         itr,
         log_format=cfg.common.log_format,
         log_file=cfg.common.log_file,
@@ -252,7 +252,6 @@ def train(
             if distributed_utils.is_master(cfg.distributed_training)
             else None
         ),
-        default_log_format="json",
         wandb_project=(
             cfg.common.wandb_project
             if distributed_utils.is_master(cfg.distributed_training)
@@ -503,7 +502,7 @@ def validate(
                 )
             )
 
-            progress = progress_bar.progress_bar(
+            progress = progress_bar.get_progress_bar(
                 itr,
                 log_format=cfg.common.log_format,
                 log_interval=cfg.common.log_interval,
@@ -514,7 +513,6 @@ def validate(
                     if distributed_utils.is_master(cfg.distributed_training)
                     else None
                 ),
-                default_log_format="json",
                 wandb_project=(
                     cfg.common.wandb_project
                     if distributed_utils.is_master(cfg.distributed_training)
